requirements.txt
streamlit
langchain
langgraph
langchain-groq
python-dotenv
PyPDF2
openai 
pytesseract 
pillow